// packages/admin/src/utils/types.ts
export interface IQueueData {
  service: string;
  status: boolean;
  queue: string;
  active: number;
  completed: number;
  delayed: number;
  failed: number;
  paused: number;
  waiting: number;
  "waiting-children": number;
}